% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckRNASeqData.R
\name{CheckRNAseqDataServer}
\alias{CheckRNAseqDataServer}
\title{Check RNASeqdata for further analysis server part}
\usage{
CheckRNAseqDataServer(
  input,
  output,
  session,
  matrix = NULL,
  annotation = NULL,
  metadata = NULL
)
}
\arguments{
\item{input, output, session}{standards \code{shiny} server arguments.}

\item{header}{Does the file have a Header}

\item{sep}{What is the file separator}
}
\description{
Check RNASeqdata for further analysis server part
}
