% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringModule.R
\name{ClusteringServer}
\alias{ClusteringServer}
\title{Draw clusters heatmap from counts matrix server side}
\usage{
ClusteringServer(
  input,
  output,
  session,
  data = NULL,
  metadata = NULL,
  printRows = FALSE
)
}
\arguments{
\item{input, output, session}{standards \code{shiny} server arguments.}

\item{data}{}

\item{name}{Does the file have a Header}
}
\description{
Draw clusters heatmap from counts matrix server side
}
