% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringModuleCNV.R
\name{ClusteringServerCNV}
\alias{ClusteringServerCNV}
\title{Draw clusters heatmap from counts matrix server side}
\usage{
ClusteringServerCNV(
  input,
  output,
  session,
  data = NULL,
  metadata = NULL,
  printRows = FALSE
)
}
\arguments{
\item{input, output, session}{standards \code{shiny} server arguments.}

\item{data}{A reactiveValues containing the table reactiveValue with table a dataframe. Example data$table  a reactive dataframe}

\item{name}{Does the file have a Header}
}
\description{
Draw clusters heatmap from counts matrix server side
}
