% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDDSModule.R
\name{CreateDdsServer}
\alias{CreateDdsServer}
\title{Creates a DDS object for further analysis server side}
\usage{
CreateDdsServer(input, output, session, countmatrix = NULL, colData = NULL)
}
\arguments{
\item{input, output, session}{standards \code{shiny} server arguments.}

\item{header}{Does the file have a Header}

\item{sep}{What is the file separator}
}
\description{
Creates a DDS object for further analysis server side
}
