% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRNAseqModeleComplex.R
\name{CreateModelServerComplex}
\alias{CreateModelServerComplex}
\title{Check RNASeqdata for further analysis server part}
\usage{
CreateModelServerComplex(
  input,
  output,
  session,
  sampleplan = NULL,
  matrix = NULL,
  batcheffect = NULL
)
}
\arguments{
\item{input, output, session}{standards \code{shiny} server arguments.}

\item{header}{Does the file have a Header}

\item{sep}{What is the file separator}
}
\description{
Check RNASeqdata for further analysis server part
}
