% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEAnalysisModule.R
\name{DEAServer}
\alias{DEAServer}
\title{Perform DE analysis server side}
\usage{
DEAServer(input, output, session, countmatrix = NULL, Model = NULL)
}
\arguments{
\item{input, output, session}{standards \code{shiny} server arguments.}

\item{header}{Does the file have a Header}

\item{sep}{What is the file separator}
}
\description{
Perform DE analysis server side
}
