% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEAnalysisModule.R, R/DEseqAnalysisModule.R
\name{DEAUI}
\alias{DEAUI}
\title{Perform DE analysis ui side}
\usage{
DEAUI(id)

DEAUI(id)
}
\arguments{
\item{id}{Module's id.}

\item{label}{Button's label.}

\item{icon}{Button's icon.}

\item{...}{Arguments passed to \code{\link{actionButton}}}
}
\value{
a \code{\link[shiny]{reactiveValues}} containing the data selected under slot \code{data}
and the name of the selected \code{data.frame} under slot \code{name}.

a \code{\link[shiny]{reactiveValues}} containing the data selected under slot \code{data}
and the name of the selected \code{data.frame} under slot \code{name}.
}
\description{



}
