% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAModule2.R
\name{DrawPCAServer2}
\alias{DrawPCAServer2}
\title{Module for drawing PCA from counts matrix server part}
\usage{
DrawPCAServer2(
  input,
  output,
  session,
  matrix = NULL,
  annotation = NULL,
  metadata = NULL
)
}
\arguments{
\item{input, output, session}{standards \code{shiny} server arguments.}

\item{header}{Does the file have a Header}

\item{sep}{What is the file separator}
}
\description{
Module for drawing PCA from counts matrix server part
}
