% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterRNAseqDataModule.R
\name{FilterRNAServer}
\alias{FilterRNAServer}
\title{Filter counts data for further analysiss server side}
\usage{
FilterRNAServer(input, output, session, data = NULL)
}
\arguments{
\item{input, output, session}{standards \code{shiny} server arguments.}

\item{data}{A reactiveValues containing the table reactiveValue with table a dataframe. Example data$table  a reactive dataframe}

\item{name}{Does the file have a Header}
}
\description{
Filter counts data for further analysiss server side
}
