% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnotationModule.R
\name{getAnnotationServer}
\alias{getAnnotationServer}
\title{Modules getting a bioMaRt annotation from genes's list server part}
\usage{
getAnnotationServer(input, output, session, matrix = NULL, annotation = NULL)
}
\arguments{
\item{input, output, session}{standards \code{shiny} server arguments.}

\item{header}{Does the file have a Header}

\item{sep}{What is the file separator}
}
\description{
Modules getting a bioMaRt annotation from genes's list server part
}
