% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChooseDataModuleFixedParameters.R
\name{ImportDataServerFixed}
\alias{ImportDataServerFixed}
\title{Module for importing text of xls files server part}
\usage{
ImportDataServerFixed(
  input,
  output,
  session,
  header = NULL,
  sep = NULL,
  ext = "txt",
  row.names = NULL
)
}
\arguments{
\item{input, output, session}{standards \code{shiny} server arguments.}

\item{header}{Does the file have a Header}

\item{sep}{What is the file separator}
}
\description{
Module for importing text of xls files server part
}
