% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChooseDataModule_User_handlable.R
\name{ImportDataServer}
\alias{ImportDataServer}
\title{Module for importing text of xls files server part}
\usage{
ImportDataServer(input, output, session)
}
\arguments{
\item{input, output, session}{standards \code{shiny} server arguments.}

\item{header}{Does the file have a Header}

\item{sep}{What is the file separator}
}
\description{
Module for importing text of xls files server part
}
