% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqnormalization.R
\name{NormalizeRNAseqDataServer}
\alias{NormalizeRNAseqDataServer}
\title{Performs normalization on RNAseq data server side}
\usage{
NormalizeRNAseqDataServer(input, output, session, matrix = NULL)
}
\arguments{
\item{matrix}{a reactive values object containing a reactiveVal table, for exemple matrix$table}
}
\description{
Performs normalization on RNAseq data server side
}
