% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqTransformations.R
\name{TransformRNAseqDataServer}
\alias{TransformRNAseqDataServer}
\title{Performs normalization on RNAseq data server part}
\usage{
TransformRNAseqDataServer(
  input,
  output,
  session,
  dds = NULL,
  minlength = NULL,
  matrix = NULL,
  gtf_path = NULL
)
}
\arguments{
\item{input, output, session}{standards \code{shiny} server arguments.}

\item{header}{Does the file have a Header}

\item{sep}{What is the file separator}
}
\description{
Performs normalization on RNAseq data server part
}
