% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqTransformations.R
\name{TransformRNAseqDataUI}
\alias{TransformRNAseqDataUI}
\title{Performs normalization on RNAseq data ui part}
\usage{
TransformRNAseqDataUI(id)
}
\arguments{
\item{id}{Module's id.}

\item{label}{Button's label.}

\item{icon}{Button's icon.}

\item{...}{Arguments passed to \code{\link{actionButton}}}
}
\value{
a \code{\link[shiny]{reactiveValues}} containing the data selected under slot \code{data}
and the name of the selected \code{data.frame} under slot \code{name}.
}
\description{

}
